Migrate to 2023.2.3 inspection engine

+++ driver/src/test/java/org/neo4j/driver/internal/retry/ExponentialBackoffRetryLogicTest.java
@@ -1267,7 +1267,7 @@ class ExponentialBackoffRetryLogicTest {
         var delay = initialDelay;
         do {
             values.add(delay);
-            delay *= multiplier;
+            delay *= (long) multiplier;
         } while (--count > 0);
         return values;
     }
